requirements.txt
fastapi
uvicorn
motor            
jinja2
python-multipart
pydantic
pydantic-settings
python-dotenv